pyproject.toml
[tool.poetry]
name = "google-ai-bot-demo"
version = "0.1.0"
description = ""
authors = ["Ian <exor2008@yandex.ru>"]
readme = "README.md"
packages = [{ include = "google_ai_bot_demo" }]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
discord-py-interactions = "^5.9.2"
bardapi = "^0.1.33"
google-generativeai = "^0.1.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 100
profile = "black"
